+++ src/main/java/com/mvc/coinsimulation/repository/postgres/custom/OrderCustomRepository.java
@@ -2,14 +2,15 @@ package com.mvc.coinsimulation.repository.postgres.custom;
 
 import com.mvc.coinsimulation.entity.Order;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 
 public interface OrderCustomRepository {
     Optional<Order> findByIdAndUserIdForUpdate(Long id, Long userId);
 
-    List<Order> findBidOrders(String code, Double price);
+    List<Order> findBidOrders(String code, BigDecimal price);
 
-    List<Order> findAskOrders(String code, Double price);
+    List<Order> findAskOrders(String code, BigDecimal price);
 
 }
